requirements.txt
streamlit
sentence-transformers
langchain
langchain-community
faiss-cpu
pymupdf
python-docx
python-pptx
pandas
openpyxl
requests
numpy
pyngrok